LayoutTests/inspector/console/command-line-api.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

var test = function()
{
    var expressions = [
        "$0.toString()",
        "$3",
        "keys([3,4]).toString()",
        "values([3,4]).toString()",
        "$('foo').toString()",
        "$$('p#foo').toString()"
    ];

    InspectorTest.selectNodeWithId("foo", step1);

    function step1(node)
    {
        if (!expressions.length) {
            InspectorTest.completeTest();
            return;
        }
        var expression = expressions[0];
        expressions.splice(0, 1);
        InspectorTest.evaluateInConsoleAndDump(expression, step1);
    }
}

</script>
</head>

<body onload="runTest()">
<p id="foo">
Tests that command line api works.
</p>

</body>
</html>
